BSR debug

+++ src/engine/bsr.h
@@ -64,7 +64,7 @@ static inline int bsr(unsigned short v) {
   unsigned short mask = 0x8000;
   for (int i = 15; i >= 0; --i) {
     if (v&mask)
-      return (int)i;
+      return i;
     mask>>=1;
   }
 
@@ -75,7 +75,7 @@ static inline int bsr32(unsigned int v) {
   unsigned int mask = 0x80000000;
   for (int i = 31; i >= 0; --i) {
     if (v&mask)
-      return (int)i;
+      return i;
     mask>>=1;
   }
